commito

+++ OoOVisual/Core/include/Core/Fetch/FetchElements.h
@@ -11,7 +11,7 @@ namespace OoOVisual
             std::unique_ptr<FrontEnd::Instruction>* instruction{};
             memory_addr_t address = 0;
             u32 timestamp = 0;
-            u32 branch_prediction = Constants::NOT_BRANCH_INSTRUCTION;
+            u8 branch_prediction = Constants::NOT_BRANCH_INSTRUCTION;
         };
         struct Fetch_Group {
             static std::vector<Fetch_Element> group;
